feat[test]support lib
